Moved telegram.js to libs and made sendMessage() available
