+++ server/libs/telegram.js
@@ -3,6 +3,6 @@ const token = 'replace me with token from database';
 const bot = new TelegramBot(token, {polling: false});
 const chatId = 'replace me with chatId from database';
 
-var sendMessage = function(message) {
+module.exports.sendMessage = function(message) {
   bot.sendMessage(chatId, message);
 };
